+++ ls_Minimap/locales/ruRU.lua
@@ -1,4 +1,4 @@
--- Contributors:
+﻿-- Contributors:
 
 local _, addon = ...
 local L = addon.L
@@ -8,7 +8,6 @@ local _G = getfenv(0)
 
 if GetLocale() ~= "ruRU" then return end
 
--- Mine
 L["AUTO_ZOOM_OUT"] = "Авто-отдаление"
 L["CHANGELOG"] = "Список изменений"
 L["CHANGELOG_FULL"] = "Полный"
@@ -16,6 +15,7 @@ L["COLLAPSE_OPTIONS"] = "Свернуть настройки"
 L["COLORED_BORDER"] = "Цветная граница"
 L["COORDS"] = "Координаты"
 L["DOWNLOADS"] = "Загрузки"
+L["EXPORT"] = "Экспорт"
 L["LINK_COPY_SUCCESS"] = "Ссылка скопирована в буфер обмена."
 L["ROUND"] = "Круглая"
 L["SHAPE"] = "Форма"