+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## Version 120005.01
+
+- Added 12.0.5 support.
+- Added profile import/export. Available at the "Profiles" tab in Blizz config panel.
+- Fixed an issue where the addon could interfere with other addons that were messing with the stopwatch.
+
 ## Version 120000.02
 
 - Added proper handling of Edit Mode layout copying, deletion, and renaming.